third_party/blink/web_tests/fast/dom/SelectorAPI/namespaced-elements-and-selectors.html

<!DOCTYPE html>
<title>querySelectorAll and namespaces</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="sandbox">
    <span></span>
</div>
<script>
setup({ single_test: true });


sandbox.appendChild(document.createElementNS("http://dummynamespace", "span"));
sandbox.appendChild(document.createElementNS("", "span"));

assert_equals(sandbox.querySelectorAll('span').length, 3);
assert_equals(sandbox.querySelectorAll('*|span').length, 3);
assert_equals(sandbox.querySelectorAll('|span').length, 1);
done();
</script>
